# Improve performance of `JohnsonSU`